fix: handle get rss feed using getStaticProps
